streamlit/requirements.txt
streamlit
torch
transformers
ultralytics
Pillow
config
